refactor(payment): Phase 9 reshape Payment.Tests into Features/
- Mirror Features/<Slice>/ layout: split PaymentEndpointsTests +
  PaymentOwnershipTests + MultiProducerWiringTests across slice folders;
  relocate command-handler tests + InternalOutboxEndpointsTests;
  move PaymentStateMachineTests to Domain/; lift cross-cutting
  HealthEndpointTests + QaSeedPresenceTests to root. Delete Api/ +
  Models/. 9 slice folders match slice count.
- Add 5 IIntegrationMap<,> unit tests (one per mapper) co-located with
  their owning slice; field-level Map preservation + DomainEventType
  assertion. Add Features/OrderCreated/OrderCreatedHandlerTests.cs.
- Pre-existing MessagingProviderBootTests RabbitMQ failure unchanged
  (eager-broker defect, blocked on shared-lib bump).
