+++ payment-microservice/Payment.Tests/Features/GetPaymentById/GetPaymentByIdOwnershipTests.cs
@@ -2,69 +2,14 @@ using System.Net;
 using System.Net.Http.Json;
 using Payment.Tests.Authentication;
 using GetByIdPaymentResponse = Payment.Service.Features.GetPaymentById.PaymentResponse;
-using GetByOrderPaymentResponse = Payment.Service.Features.GetPaymentByOrder.PaymentResponse;
 
-namespace Payment.Tests.Api;
+namespace Payment.Tests.Features.GetPaymentById;
 
-public class PaymentOwnershipTests : IntegrationTestBase
+public class GetPaymentByIdOwnershipTests : IntegrationTestBase
 {
-    public PaymentOwnershipTests(PaymentWebApplicationFactory webApplicationFactory)
+    public GetPaymentByIdOwnershipTests(PaymentWebApplicationFactory webApplicationFactory)
         : base(webApplicationFactory)
     {
-    }
-
-    [Fact]
-    public async Task GetByOrder_WhenCustomerOwnsPayment_ReturnsOk()
-    {
-        var (orderId, paymentId, customerId) = await SeedPaymentAsync();
-
-        var response = await CreateCustomerClient(customerId).GetAsync($"/by-order/{orderId}");
-
-        response.EnsureSuccessStatusCode();
-        var body = await response.Content.ReadFromJsonAsync<GetByOrderPaymentResponse>();
-        Assert.NotNull(body);
-        Assert.Equal(paymentId, body.PaymentId);
-        Assert.Equal(orderId, body.OrderId);
-        Assert.Equal(customerId, body.CustomerId);
-    }
-
-    [Fact]
-    public async Task GetByOrder_WhenCustomerIsNotOwner_ReturnsNotFound()
-    {
-        var (orderId, _, _) = await SeedPaymentAsync();
-
-        var response = await CreateCustomerClient("different-customer").GetAsync($"/by-order/{orderId}");
-
-        Assert.Equal(HttpStatusCode.NotFound, response.StatusCode);
-    }
-
-    [Fact]
-    public async Task GetByOrder_WhenAdmin_ReturnsOkRegardlessOfOwnership()
-    {
-        var (orderId, paymentId, _) = await SeedPaymentAsync();
-
-        var response = await CreateAuthenticatedClient().GetAsync($"/by-order/{orderId}");
-
-        response.EnsureSuccessStatusCode();
-        var body = await response.Content.ReadFromJsonAsync<GetByOrderPaymentResponse>();
-        Assert.NotNull(body);
-        Assert.Equal(paymentId, body.PaymentId);
-    }
-
-    [Fact]
-    public async Task GetByOrder_WhenUnauthenticated_ReturnsUnauthorized()
-    {
-        var response = await HttpClient.GetAsync($"/by-order/{Guid.NewGuid()}");
-
-        Assert.Equal(HttpStatusCode.Unauthorized, response.StatusCode);
-    }
-
-    [Fact]
-    public async Task GetByOrder_WhenNotFound_ReturnsNotFound()
-    {
-        var response = await CreateAuthenticatedClient().GetAsync($"/by-order/{Guid.NewGuid()}");
-
-        Assert.Equal(HttpStatusCode.NotFound, response.StatusCode);
     }
 
     [Fact]
